[SIEM] Refactor IP details page (#47593)
