+++ x-pack/legacy/plugins/siem/public/components/link_to/redirect_to_hosts.tsx
@@ -9,6 +9,7 @@ import { RouteComponentProps } from 'react-router-dom';
 
 import { RedirectWrapper } from './redirect_wrapper';
 import { HostsTableType } from '../../store/hosts/model';
+import { SiemPageName } from '../../pages/home/types';
 
 export type HostComponentProps = RouteComponentProps<{
   detailName: string;
@@ -24,7 +25,7 @@ export const RedirectToHostsPage = ({
 }: HostComponentProps) => {
   const defaultSelectedTab = HostsTableType.hosts;
   const selectedTab = tabName ? tabName : defaultSelectedTab;
-  const to = `/hosts/${selectedTab}${search}`;
+  const to = `/${SiemPageName.hosts}/${selectedTab}${search}`;
 
   return <RedirectWrapper to={to} />;
 };
@@ -37,16 +38,18 @@ export const RedirectToHostDetailsPage = ({
 }: HostComponentProps) => {
   const defaultSelectedTab = HostsTableType.authentications;
   const selectedTab = tabName ? tabName : defaultSelectedTab;
-  const to = `/hosts/${detailName}/${selectedTab}${search}`;
+  const to = `/${SiemPageName.hosts}/${detailName}/${selectedTab}${search}`;
   return <RedirectWrapper to={to} />;
 };
 
-export const getHostsUrl = () => '#/link-to/hosts';
+const baseHostsUrl = `#/link-to/${SiemPageName.hosts}`;
 
-export const getTabsOnHostsUrl = (tabName: HostsTableType) => `#/link-to/hosts/${tabName}`;
+export const getHostsUrl = () => baseHostsUrl;
 
-export const getHostDetailsUrl = (detailName: string) => `#/link-to/hosts/${detailName}`;
+export const getTabsOnHostsUrl = (tabName: HostsTableType) => `${baseHostsUrl}/${tabName}`;
+
+export const getHostDetailsUrl = (detailName: string) => `${baseHostsUrl}/${detailName}`;
 
 export const getTabsOnHostDetailsUrl = (detailName: string, tabName: HostsTableType) => {
-  return `#/link-to/hosts/${detailName}/${tabName}`;
+  return `${baseHostsUrl}/${detailName}/${tabName}`;
 };
