+++ x-pack/legacy/plugins/siem/public/components/link_to/redirect_to_overview.tsx
@@ -7,13 +7,14 @@
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
 import { RedirectWrapper } from './redirect_wrapper';
+import { SiemPageName } from '../../pages/home/types';
 
 export type OverviewComponentProps = RouteComponentProps<{
   search: string;
 }>;
 
 export const RedirectToOverviewPage = ({ location: { search } }: OverviewComponentProps) => (
-  <RedirectWrapper to={`/overview${search}`} />
+  <RedirectWrapper to={`/${SiemPageName.overview}${search}`} />
 );
 
-export const getOverviewUrl = () => '#/link-to/overview';
+export const getOverviewUrl = () => `#/link-to/${SiemPageName.overview}`;
