+++ x-pack/legacy/plugins/siem/public/components/link_to/redirect_to_timelines.tsx
@@ -7,15 +7,14 @@
 import React from 'react';
 import { RouteComponentProps } from 'react-router';
 import { RedirectWrapper } from './redirect_wrapper';
+import { SiemPageName } from '../../pages/home/types';
 
 export type TimelineComponentProps = RouteComponentProps<{
   search: string;
 }>;
 
-export const TIMELINES_PAGE_NAME = 'timelines';
-
 export const RedirectToTimelinesPage = ({ location: { search } }: TimelineComponentProps) => (
-  <RedirectWrapper to={`/${TIMELINES_PAGE_NAME}${search}`} />
+  <RedirectWrapper to={`/${SiemPageName.timelines}${search}`} />
 );
 
-export const getTimelinesUrl = () => `#/link-to/${TIMELINES_PAGE_NAME}`;
+export const getTimelinesUrl = () => `#/link-to/${SiemPageName.timelines}`;
