+++ x-pack/legacy/plugins/siem/public/components/navigation/tab_navigation/index.test.tsx
@@ -9,7 +9,8 @@ import * as React from 'react';
 
 import { TabNavigation } from './';
 import { TabNavigationProps } from './types';
-import { navTabs, SiemPageName } from '../../../pages/home/home_navigations';
+import { navTabs } from '../../../pages/home/home_navigations';
+import { SiemPageName } from '../../../pages/home/types';
 import { HostsTableType } from '../../../store/hosts/model';
 import { navTabsHostDetails } from '../../../pages/hosts/details/nav_tabs';
 import { CONSTANTS } from '../../url_state/constants';
