+++ x-pack/legacy/plugins/siem/public/components/url_state/helpers.test.ts
@@ -3,7 +3,8 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-import { navTabs, SiemPageName } from '../../pages/home/home_navigations';
+import { navTabs } from '../../pages/home/home_navigations';
+import { SiemPageName } from '../../pages/home/types';
 import { isKqlForRoute, getTitle } from './helpers';
 import { CONSTANTS } from './constants';
 import { HostsType } from '../../store/hosts/model';
