+++ x-pack/legacy/plugins/siem/public/components/url_state/index_mocked.test.tsx
@@ -8,7 +8,7 @@ import { mount } from 'enzyme';
 import * as React from 'react';
 
 import { HookWrapper } from '../../mock/hook_wrapper';
-import { SiemPageName } from '../../pages/home/home_navigations';
+import { SiemPageName } from '../../pages/home/types';
 
 import { CONSTANTS } from './constants';
 import { getFilterQuery, getMockPropsObj, mockHistory, testCases } from './test_dependencies';
