+++ x-pack/legacy/plugins/siem/public/pages/home/home_navigations.tsx
@@ -4,28 +4,13 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 import * as i18n from './translations';
+import { SiemPageName, SiemNavTab } from './types';
 import {
   getOverviewUrl,
   getNetworkUrl,
   getTimelinesUrl,
   getHostsUrl,
 } from '../../components/link_to';
-import { NavTab } from '../../components/navigation/types';
-
-export enum SiemPageName {
-  overview = 'overview',
-  hosts = 'hosts',
-  network = 'network',
-  timelines = 'timelines',
-}
-
-export type SiemNavTabKey =
-  | SiemPageName.overview
-  | SiemPageName.hosts
-  | SiemPageName.network
-  | SiemPageName.timelines;
-
-export type SiemNavTab = Record<SiemNavTabKey, NavTab>;
 
 export const navTabs: SiemNavTab = {
   [SiemPageName.overview]: {
