+++ x-pack/legacy/plugins/siem/public/pages/home/index.tsx
@@ -29,6 +29,7 @@ import { MlPopover } from '../../components/ml_popover/ml_popover';
 import { MlHostConditionalContainer } from '../../components/ml/conditional_links/ml_host_conditional_container';
 import { MlNetworkConditionalContainer } from '../../components/ml/conditional_links/ml_network_conditional_container';
 import { navTabs } from './home_navigations';
+import { SiemPageName } from './types';
 import { UseUrlState } from '../../components/url_state';
 import { SpyRoute } from '../../utils/route/spy_routes';
 
@@ -137,21 +138,27 @@ export const HomePage = pure(() => (
                     </EuiFlexGroup>
                   </NavGlobal>
                   <Switch>
-                    <Redirect from="/" exact={true} to="/overview" />
-                    <Route path="/:pageName(overview)" render={() => <Overview />} />
+                    <Redirect from="/" exact={true} to={`/${SiemPageName.overview}`} />
                     <Route
-                      path="/:pageName(hosts)"
+                      path={`/:pageName(${SiemPageName.overview})`}
+                      render={() => <Overview />}
+                    />
+                    <Route
+                      path={`/:pageName(${SiemPageName.hosts})`}
                       render={({ match, location }) => (
                         <HostsContainer url={match.url} location={location} />
                       )}
                     />
                     <Route
-                      path="/:pageName(network)"
+                      path={`/:pageName(${SiemPageName.network})`}
                       render={({ match, location }) => (
                         <NetworkContainer url={match.url} location={location} />
                       )}
                     />
-                    <Route path="/:pageName(timelines)" render={() => <Timelines />} />
+                    <Route
+                      path={`/:pageName(${SiemPageName.timelines})`}
+                      render={() => <Timelines />}
+                    />
                     <Route path="/link-to" component={LinkToPage} />
                     <Route
                       path="/ml-hosts"
