+++ x-pack/legacy/plugins/siem/public/pages/hosts/index.tsx
@@ -18,9 +18,10 @@ import {
 import { HostsBody } from './hosts_body';
 import { HostsTableType } from '../../store/hosts/model';
 import { GlobalTime } from '../../containers/global_time';
+import { SiemPageName } from '../home/types';
 import { Hosts } from './hosts';
 
-const hostsPagePath = `/:pageName(hosts)`;
+const hostsPagePath = `/:pageName(${SiemPageName.hosts})`;
 
 const getHostsTabPath = (pagePath: string) =>
   `${pagePath}/:tabName(` +
@@ -236,9 +237,12 @@ export const HostsContainer = React.memo<Props>(({ url }) => (
           )}
         />
         <Route
-          path="/hosts/"
+          path={`/${SiemPageName.hosts}/`}
           render={({ location: { search = '' } }) => (
-            <Redirect from={`/hosts/"`} to={`/hosts/${HostsTableType.hosts}${search}`} />
+            <Redirect
+              from={`/${SiemPageName.hosts}/"`}
+              to={`/${SiemPageName.hosts}/${HostsTableType.hosts}${search}`}
+            />
           )}
         />
       </Switch>
